# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      responses:
        "200":
          description: A list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by id
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: A job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      responses:
        "200":
          description: A list of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all catalog products
      responses:
        "200":
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a catalog product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a catalog product
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a catalog product
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    JobStatus:
      type: string
      enum:
        - New Lead
        - Estimate Scheduled
        - Estimate Completed
        - Proposal Sent
        - Approved
        - Material Ordered
        - Material Received
        - Scheduled
        - In Progress
        - Final Walkthrough
        - Completed
        - Invoiced
    PriorityLevel:
      type: string
      enum: [Low, Medium, High]
    RiskLevel:
      type: string
      enum: [Low, Medium, High]
    CrewAssigned:
      type: string
      enum: [Crew A, Crew B, Crew C, Crew D, Unassigned]
    FlooringType:
      type: string
      enum:
        - Carpet
        - Hardwood
        - Tile
        - Laminate
        - Luxury Vinyl Plank (LVP)
        - Luxury Vinyl Tile (LVT)
        - Waterproof Flooring
        - Commercial Carpet Tile
        - Commercial LVT
    MaterialStatusBadge:
      type: string
      enum:
        - Ordered
        - In Transit
        - Received
        - Delayed
        - Damaged
        - Missing Items
...
    InvoiceStatus:
      type: string
      enum: [Draft, Sent, Paid, Overdue]
    InvoiceLineCategory:
      type: string
      enum: [Labor, Materials, Add-on]
    ProductCategory:
      type: string
      enum:
        - Vinyl Plank
        - Vinyl Tile
        - Carpet
        - Hardwood
        - Tile
        - Laminate
        - Waterproof
        - Commercial
        - Trim/Supplies
    ProductUnit:
      type: string
      enum:
        - sqft
        - box
        - lineal ft
        - each
    ProductInventoryType:
      type: string
      enum:
        - Inventory
        - Special Order
...
    Room:
      type: object
      required: [id, name, length, width]
      properties:
        id:
          type: string
        name:
          type: string
        length:
          type: number
        width:
          type: number
        scope:
          type: string
    StageEvent:
      type: object
      required: [stage, at]
      properties:
        stage:
          $ref: "#/components/schemas/JobStatus"
        at:
          type: string
...
    Job:
      type: object
      required:
        - id
        - jobNumber
        - customerName
        - phone
        - email
        - address
        - city
        - flooringType
        - rooms
        - scopeOfWork
        - squareFootage
        - crewAssigned
        - materialStatus
        - laborEstimate
        - estLaborHours
        - materialEstimate
        - estRevenue
        - estGrossProfit
        - grossMarginPct
        - notes
        - priorityLevel
        - riskLevel
        - status
        - stageHistory
        - shareToken
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        jobNumber:
          type: string
        customerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        scopeOfWork:
          type: string
        squareFootage:
          type: number
        crewAssigned:
          $ref: "#/components/schemas/CrewAssigned"
        salespersonId:
          type: string
          nullable: true
        estStartDate:
          type: string
        estCompletionDate:
          type: string
        materialStatus:
          $ref: "#/components/schemas/MaterialStatusBadge"
        laborEstimate:
          type: number
        estLaborHours:
          type: number
        materialEstimate:
          type: number
        estRevenue:
          type: number
        estGrossProfit:
          type: number
        grossMarginPct:
          type: number
        actualRevenue:
          type: number
        actualLaborCost:
          type: number
        actualMaterialCost:
          type: number
        actualAddOnCost:
          type: number
        notes:
          type: string
        priorityLevel:
          $ref: "#/components/schemas/PriorityLevel"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        status:
          $ref: "#/components/schemas/JobStatus"
        stageHistory:
          type: array
          items:
            $ref: "#/components/schemas/StageEvent"
        shareToken:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    JobInput:
      type: object
      required:
        - customerName
        - phone
        - email
        - address
        - city
        - flooringType
        - rooms
        - squareFootage
        - crewAssigned
        - materialStatus
        - laborEstimate
        - materialEstimate
        - estRevenue
        - estGrossProfit
        - grossMarginPct
        - notes
        - priorityLevel
        - riskLevel
        - status
      properties:
        customerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        scopeOfWork:
          type: string
        squareFootage:
          type: number
        crewAssigned:
          $ref: "#/components/schemas/CrewAssigned"
        salespersonId:
          type: string
          nullable: true
        estStartDate:
          type: string
        estCompletionDate:
          type: string
        materialStatus:
          $ref: "#/components/schemas/MaterialStatusBadge"
        laborEstimate:
          type: number
        estLaborHours:
          type: number
        materialEstimate:
          type: number
        estRevenue:
          type: number
        estGrossProfit:
          type: number
        grossMarginPct:
          type: number
        actualRevenue:
          type: number
        actualLaborCost:
          type: number
        actualMaterialCost:
          type: number
        actualAddOnCost:
          type: number
        notes:
          type: string
        priorityLevel:
          $ref: "#/components/schemas/PriorityLevel"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        status:
          $ref: "#/components/schemas/JobStatus"
    JobUpdate:
      type: object
      properties:
        customerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        scopeOfWork:
          type: string
        squareFootage:
          type: number
        crewAssigned:
          $ref: "#/components/schemas/CrewAssigned"
        salespersonId:
          type: string
          nullable: true
        estStartDate:
          type: string
        estCompletionDate:
          type: string
        materialStatus:
          $ref: "#/components/schemas/MaterialStatusBadge"
        laborEstimate:
          type: number
        estLaborHours:
          type: number
        materialEstimate:
          type: number
        estRevenue:
          type: number
        estGrossProfit:
          type: number
        grossMarginPct:
          type: number
        actualRevenue:
          type: number
        actualLaborCost:
          type: number
        actualMaterialCost:
          type: number
        actualAddOnCost:
          type: number
        notes:
          type: string
        priorityLevel:
          $ref: "#/components/schemas/PriorityLevel"
        riskLevel:
          $ref: "#/components/schemas/RiskLevel"
        status:
          $ref: "#/components/schemas/JobStatus"
...
    InvoiceLineItem:
      type: object
      required: [id, description, category, quantity, unitPrice]
      properties:
        id:
          type: string
        description:
          type: string
        category:
          $ref: "#/components/schemas/InvoiceLineCategory"
        quantity:
          type: number
        unitPrice:
          type: number
    Invoice:
      type: object
      required:
        - id
        - invoiceNumber
        - jobId
        - jobNumber
        - customerName
        - lineItems
        - subtotal
        - taxableAmount
        - taxAmount
        - discountAmount
        - total
        - depositAmount
        - paidAmount
        - balanceAmount
        - refundedAmount
        - taxCode
        - paymentReference
        - status
        - issueDate
        - dueDate
        - notes
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        jobId:
          type: string
        jobNumber:
          type: string
        customerName:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        taxableAmount:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        depositAmount:
          type: number
        paidAmount:
          type: number
        balanceAmount:
          type: number
        refundedAmount:
          type: number
        taxCode:
          type: string
        paymentReference:
          type: string
        paidAt:
          type: string
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        issueDate:
          type: string
        dueDate:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    InvoiceInput:
      type: object
      required:
        - jobId
        - customerName
        - lineItems
        - status
      properties:
        jobId:
          type: string
        jobNumber:
          type: string
        customerName:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        depositAmount:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        taxCode:
          type: string
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        issueDate:
          type: string
        dueDate:
          type: string
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        customerName:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        depositAmount:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        taxCode:
          type: string
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        issueDate:
          type: string
        dueDate:
          type: string
        notes:
          type: string
    Product:
      type: object
      required:
        - id
        - name
        - category
        - sku
        - supplier
        - color
        - unit
        - cost
        - price
        - quantityOnHand
        - inventoryType
        - active
        - notes
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        sku:
          type: string
        supplier:
          type: string
        color:
          type: string
        unit:
          $ref: "#/components/schemas/ProductUnit"
        cost:
          type: number
        price:
          type: number
        quantityOnHand:
          type: number
        inventoryType:
          $ref: "#/components/schemas/ProductInventoryType"
        active:
          type: boolean
        notes:
          type: string
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
    ProductInput:
      type: object
      required:
        - name
        - category
        - unit
      properties:
        name:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        sku:
          type: string
        supplier:
          type: string
        color:
          type: string
        unit:
          $ref: "#/components/schemas/ProductUnit"
        cost:
          type: number
        price:
          type: number
        quantityOnHand:
          type: number
        inventoryType:
          $ref: "#/components/schemas/ProductInventoryType"
        active:
          type: boolean
        notes:
          type: string
        imageUrl:
          type: string
          nullable: true
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        sku:
          type: string
        supplier:
          type: string
        color:
          type: string
        unit:
          $ref: "#/components/schemas/ProductUnit"
        cost:
          type: number
        price:
          type: number
        quantityOnHand:
          type: number
        inventoryType:
          $ref: "#/components/schemas/ProductInventoryType"
        active:
          type: boolean
        notes:
          type: string
        imageUrl:
          type: string
          nullable: true
...