# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceLineItem } from './invoiceLineItem';
import type { InvoiceStatus } from './invoiceStatus';

export interface Invoice {
  id: string;
  invoiceNumber: string;
  jobId: string;
  jobNumber: string;
  customerName: string;
  lineItems: InvoiceLineItem[];
  subtotal: number;
  taxableAmount: number;
  taxAmount: number;
  discountAmount: number;
  total: number;
  depositAmount: number;
  paidAmount: number;
  balanceAmount: number;
  refundedAmount: number;
  taxCode: string;
  paymentReference: string;
  paidAt?: string;
  status: InvoiceStatus;
  issueDate: string;
  dueDate: string;
  notes: string;
  createdAt: string;
  updatedAt: string;
}