# lib/api-zod/src/generated/types/invoiceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceLineItem } from './invoiceLineItem';
import type { InvoiceStatus } from './invoiceStatus';

export interface InvoiceUpdate {
  customerName?: string;
  lineItems?: InvoiceLineItem[];
  depositAmount?: number;
  taxAmount?: number;
  discountAmount?: number;
  taxCode?: string;
  status?: InvoiceStatus;
  issueDate?: string;
  dueDate?: string;
  notes?: string;
}